AI-finance-claude-ai-finance-crypto-tools-sZV7h/AI-finance-claude-ai-finance-crypto-tools-sZV7h/requirements.txt
anthropic>=0.40.0
ccxt>=4.3.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
rich>=13.7.0
requests>=2.31.0
pyyaml>=6.0.0
